// src/types.ts
export type SQLiteType = 'BLOB' | 'INTEGER' | 'NULL' | 'REAL' | 'TEXT'

/**
 * SQLite supported column types.
 *
 * These types are supported by SQLite but not as native storage classes.
 * They are often used as aliases for other types or for specific purposes.
 * @see https://www.sqlite.org/datatype3.html#type_affinity
 */
export type SQLiteSupportType = 'BIGINT' | 'BOOLEAN' | 'DATE' | 'DATETIME' | 'FLOAT' | 'NUMERIC'